+++ src/menu/lefthandmenu.cpp
@@ -139,9 +139,6 @@ void LeftHandMenu::resize() {
 
     mSpacer->setGeometry(0, 0, this->width(), height());
 
-    mLightsButton->setGeometry(0, yPos, this->width(), buttonHeight);
-    yPos += mLightsButton->height();
-
     mSingleColorButton->setGeometry(0, yPos, this->width(), buttonHeight);
     yPos += mSingleColorButton->height();
 
@@ -150,6 +147,9 @@ void LeftHandMenu::resize() {
 
     mMoodButton->setGeometry(0, yPos, this->width(), buttonHeight);
     yPos += mMoodButton->height();
+
+    mLightsButton->setGeometry(0, yPos, this->width(), buttonHeight);
+    yPos += mLightsButton->height();
 
     mTimeoutButton->setGeometry(0, yPos, this->width(), buttonHeight);
     yPos += mTimeoutButton->height();
@@ -163,7 +163,7 @@ void LeftHandMenu::resize() {
 
 #ifndef DISABLE_LIGHTS_MENU
     QRect lightMenuRect(0, yPos, this->width(), this->height() - yPos);
-    mLightMenu->resize(lightMenuRect, buttonHeight);
+    mLightMenu->resize(lightMenuRect, buttonHeight * 1.2);
 #endif
 }